Change to just demonstrate Zig code
I'm moving my blog to be in the same (private) repo as the rest
of my website. Some people have starred the blog repo after I linked
to it in a lobste.rs comment about the Zig markdown renderer. So
I'm planning to transition this repo into a library or a group of
libraries for rendering template, markdown, MathML, etc.